tests/requirements.txt
# Packages needed to run the tests.
# Switch into a virtual environment
# pip install -r requirements.txt

dash[dev,testing]>=2.7.0
werkzeug

# Used by test_states
pillow>=4
numpy
scikit-image>=0.16

# Required by testing
chromedriver-binary-auto  # Needs to be installed after installing the browser.
pytest
pytest-base-url
pytest-flask
pytest-html
pytest-metadata
py
pytest-selenium
pytest-variables
pytest-xdist
